requirement.txt
openai-whisper
tqdm
pyyaml
